fix some warm

+++ NACommon/NACommonFoundation/presentation/extend/view/plugin/XXJInputAlertView/UIAlertView+Input.h
@@ -8,36 +8,39 @@
 
 #import <UIKit/UIKit.h>
 
-typedef void (^DismissTextInputBLock)(UITextField *tf, int buttonIndex);
-typedef void (^DismissBlock)(int buttonIndex);
+typedef void (^DismissTextInputBLock)(UITextField *tf, long buttonIndex);
+typedef void (^DismissBlock)(long buttonIndex);
 typedef void (^CancelBlock)();
 
 @interface UIAlertView (Input)
 
-+ (UIAlertView *)alertViewStylePlainTextInputWithTitle:(NSString *)title
-                                               message:(NSString *)message
-                                          defaultValue:(NSString *)defaultValue
-                                     cancelButtonTitle:(NSString *)cancelButtonTitle
-                                     otherButtonTitles:(NSArray *)otherButtons
-                                             onDismiss:(DismissTextInputBLock)dismissed
-                                              onCancel:(CancelBlock)cancelled;
++ (UIAlertView *)
+alertViewStylePlainTextInputWithTitle:(NSString *)title
+                              message:(NSString *)message
+                         defaultValue:(NSString *)defaultValue
+                    cancelButtonTitle:(NSString *)cancelButtonTitle
+                    otherButtonTitles:(NSArray *)otherButtons
+                            onDismiss:(DismissTextInputBLock)dismissed
+                             onCancel:(CancelBlock)cancelled;
 
-+ (UIAlertView *)alertViewStylePlainTextInputWithTitle:(NSString *)title
-                                               message:(NSString *)message
-                                           placeholder:(NSString *)placehoder
-                                          keyboardType:(UIKeyboardType)keyboardType
-                                     cancelButtonTitle:(NSString *)cancelButtonTitle
-                                     otherButtonTitles:(NSArray *)otherButtons
-                                             onDismiss:(DismissTextInputBLock)dismissed
-                                              onCancel:(CancelBlock)cancelled;
++ (UIAlertView *)
+alertViewStylePlainTextInputWithTitle:(NSString *)title
+                              message:(NSString *)message
+                          placeholder:(NSString *)placehoder
+                         keyboardType:(UIKeyboardType)keyboardType
+                    cancelButtonTitle:(NSString *)cancelButtonTitle
+                    otherButtonTitles:(NSArray *)otherButtons
+                            onDismiss:(DismissTextInputBLock)dismissed
+                             onCancel:(CancelBlock)cancelled;
 
-+ (UIAlertView *)alertViewStylePlainTextInputWithTitle:(NSString *)title
-                                               message:(NSString *)message
-                                          defaultValue:(NSString *)defaultValue
-                                          keyboardType:(UIKeyboardType)keyboardType
-                                     cancelButtonTitle:(NSString *)cancelButtonTitle
-                                     otherButtonTitles:(NSArray *)otherButtons
-                                             onDismiss:(DismissTextInputBLock)dismissed
-                                              onCancel:(CancelBlock)cancelled;
++ (UIAlertView *)
+alertViewStylePlainTextInputWithTitle:(NSString *)title
+                              message:(NSString *)message
+                         defaultValue:(NSString *)defaultValue
+                         keyboardType:(UIKeyboardType)keyboardType
+                    cancelButtonTitle:(NSString *)cancelButtonTitle
+                    otherButtonTitles:(NSArray *)otherButtons
+                            onDismiss:(DismissTextInputBLock)dismissed
+                             onCancel:(CancelBlock)cancelled;
 
 @end
